📄 Update tracing example for new backtracking tag
It’s easier to generalise tracing contexts if tags either have no attributes or a single “name” attribute. Because of this, <if> and <try> have been changed to remove the backtracking attribute, and instead a <backtracking> context has been added which tells the editor that everything inside that context will be reversed.